Add new feature to server: unique citynames.
Activated through the "set citynames X" server option. If set to 0,
there are no restrictions. If set to 1 (new default), city names have
to be unique to a player. If set to 2 or 3, city names have to be
globally unique. If set to 3, a player isn't allowed to use a default
city name of another nations and it must also be globally unique.

Patch by tkurppa <teemu@hoasnet.fi>

[[originally from CVS; cvs2svn created svn r4996]]

+++ server/unithand.c
@@ -544,6 +544,11 @@ static void city_build(struct player *pplayer, struct unit *punit,
 		       "such a stupid name."));
     return;
   }
+
+  if (!is_allowed_city_name(pplayer, city_name, punit->x, punit->y, TRUE)) {
+    return;
+  }
+
   create_city(pplayer, punit->x, punit->y, city_name);
   wipe_unit(punit);
 }
